Crate rusoto_budgets

All public APIs for AWS Budgets

If you're using the service, you're probably looking for BudgetsClient and Budgets.

Structs

Budget
AWS Budget model

BudgetsClient
A client for the AWSBudgets API.

CalculatedSpend
A structure holds the actual and forecasted spend for a budget.

CostTypes
This includes the options for getting the cost of a budget.

CreateBudgetRequest
Request of CreateBudget

CreateBudgetResponse
Response of CreateBudget

CreateNotificationRequest
Request of CreateNotification

CreateNotificationResponse
Response of CreateNotification

CreateSubscriberRequest
Request of CreateSubscriber

CreateSubscriberResponse
Response of CreateSubscriber

DeleteBudgetRequest
Request of DeleteBudget

DeleteBudgetResponse
Response of DeleteBudget

DeleteNotificationRequest
Request of DeleteNotification

DeleteNotificationResponse
Response of DeleteNotification

DeleteSubscriberRequest
Request of DeleteSubscriber

DeleteSubscriberResponse
Response of DeleteSubscriber

DescribeBudgetRequest
Request of DescribeBudget

DescribeBudgetResponse
Response of DescribeBudget

DescribeBudgetsRequest
Request of DescribeBudgets

DescribeBudgetsResponse
Response of DescribeBudgets

DescribeNotificationsForBudgetRequest
Request of DescribeNotificationsForBudget

DescribeNotificationsForBudgetResponse
Response of GetNotificationsForBudget

DescribeSubscribersForNotificationRequest
Request of DescribeSubscribersForNotification

DescribeSubscribersForNotificationResponse
Response of DescribeSubscribersForNotification

Notification
Notification model. Each budget may contain multiple notifications with different settings.

NotificationWithSubscribers
A structure to relate notification and a list of subscribers who belong to the notification.

Spend
A structure represent either a cost spend or usage spend. Contains an amount and a unit.

Subscriber
Subscriber model. Each notification may contain multiple subscribers with different addresses.

TimePeriod
A time period indicated the start date and end date of a budget.

UpdateBudgetRequest
Request of UpdateBudget

UpdateBudgetResponse
Response of UpdateBudget

UpdateNotificationRequest
Request of UpdateNotification

UpdateNotificationResponse
Response of UpdateNotification

UpdateSubscriberRequest
Request of UpdateSubscriber

UpdateSubscriberResponse
Response of UpdateSubscriber

Enums

CreateBudgetError
Errors returned by CreateBudget

CreateNotificationError
Errors returned by CreateNotification

CreateSubscriberError
Errors returned by CreateSubscriber

DeleteBudgetError
Errors returned by DeleteBudget

DeleteNotificationError
Errors returned by DeleteNotification

DeleteSubscriberError
Errors returned by DeleteSubscriber

DescribeBudgetError
Errors returned by DescribeBudget

DescribeBudgetsError
Errors returned by DescribeBudgets

DescribeNotificationsForBudgetError
Errors returned by DescribeNotificationsForBudget

DescribeSubscribersForNotificationError
Errors returned by DescribeSubscribersForNotification

UpdateBudgetError
Errors returned by UpdateBudget

UpdateNotificationError
Errors returned by UpdateNotification

UpdateSubscriberError
Errors returned by UpdateSubscriber

Traits

Budgets
Trait representing the capabilities of the AWSBudgets API. AWSBudgets clients implement this trait.