Struct rusoto_budgets::BudgetsClient

pub struct BudgetsClient { /* fields omitted */ }

A client for the AWSBudgets API.

Methods

impl BudgetsClient

pub fn new(region: Region) -> BudgetsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> BudgetsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,