Trait rusoto_budgets::Budgets

pub trait Budgets {
    fn create_budget(
        &self, 
        input: CreateBudgetRequest
    ) -> RusotoFuture<CreateBudgetResponse, CreateBudgetError>;

    fn create_notification(
        &self, 
        input: CreateNotificationRequest
    ) -> RusotoFuture<CreateNotificationResponse, CreateNotificationError>;

    fn create_subscriber(
        &self, 
        input: CreateSubscriberRequest
    ) -> RusotoFuture<CreateSubscriberResponse, CreateSubscriberError>;

    fn delete_budget(
        &self, 
        input: DeleteBudgetRequest
    ) -> RusotoFuture<DeleteBudgetResponse, DeleteBudgetError>;

    fn delete_notification(
        &self, 
        input: DeleteNotificationRequest
    ) -> RusotoFuture<DeleteNotificationResponse, DeleteNotificationError>;

    fn delete_subscriber(
        &self, 
        input: DeleteSubscriberRequest
    ) -> RusotoFuture<DeleteSubscriberResponse, DeleteSubscriberError>;

    fn describe_budget(
        &self, 
        input: DescribeBudgetRequest
    ) -> RusotoFuture<DescribeBudgetResponse, DescribeBudgetError>;

    fn describe_budget_performance_history(
        &self, 
        input: DescribeBudgetPerformanceHistoryRequest
    ) -> RusotoFuture<DescribeBudgetPerformanceHistoryResponse, DescribeBudgetPerformanceHistoryError>;

    fn describe_budgets(
        &self, 
        input: DescribeBudgetsRequest
    ) -> RusotoFuture<DescribeBudgetsResponse, DescribeBudgetsError>;

    fn describe_notifications_for_budget(
        &self, 
        input: DescribeNotificationsForBudgetRequest
    ) -> RusotoFuture<DescribeNotificationsForBudgetResponse, DescribeNotificationsForBudgetError>;

    fn describe_subscribers_for_notification(
        &self, 
        input: DescribeSubscribersForNotificationRequest
    ) -> RusotoFuture<DescribeSubscribersForNotificationResponse, DescribeSubscribersForNotificationError>;

    fn update_budget(
        &self, 
        input: UpdateBudgetRequest
    ) -> RusotoFuture<UpdateBudgetResponse, UpdateBudgetError>;

    fn update_notification(
        &self, 
        input: UpdateNotificationRequest
    ) -> RusotoFuture<UpdateNotificationResponse, UpdateNotificationError>;

    fn update_subscriber(
        &self, 
        input: UpdateSubscriberRequest
    ) -> RusotoFuture<UpdateSubscriberResponse, UpdateSubscriberError>;
}

Trait representing the capabilities of the AWSBudgets API. AWSBudgets clients implement this trait.

Required methods

fn create_budget(
&self,
input: CreateBudgetRequest
) -> RusotoFuture<CreateBudgetResponse, CreateBudgetError>

Creates a budget and, if included, notifications and subscribers.

Only one of BudgetLimit or PlannedBudgetLimits can be present in the syntax at one time. Use the syntax that matches your case. The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

fn create_notification(
&self,
input: CreateNotificationRequest
) -> RusotoFuture<CreateNotificationResponse, CreateNotificationError>

Creates a notification. You must create the budget before you create the associated notification.

fn create_subscriber(
&self,
input: CreateSubscriberRequest
) -> RusotoFuture<CreateSubscriberResponse, CreateSubscriberError>

Creates a subscriber. You must create the associated budget and notification before you create the subscriber.

fn delete_budget(
&self,
input: DeleteBudgetRequest
) -> RusotoFuture<DeleteBudgetResponse, DeleteBudgetError>

Deletes a budget. You can delete your budget at any time.

Deleting a budget also deletes the notifications and subscribers that are associated with that budget.

fn delete_notification(
&self,
input: DeleteNotificationRequest
) -> RusotoFuture<DeleteNotificationResponse, DeleteNotificationError>

Deletes a notification.

Deleting a notification also deletes the subscribers that are associated with the notification.

fn delete_subscriber(
&self,
input: DeleteSubscriberRequest
) -> RusotoFuture<DeleteSubscriberResponse, DeleteSubscriberError>

Deletes a subscriber.

Deleting the last subscriber to a notification also deletes the notification.

fn describe_budget(
&self,
input: DescribeBudgetRequest
) -> RusotoFuture<DescribeBudgetResponse, DescribeBudgetError>

Describes a budget.

The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

fn describe_budget_performance_history(
&self,
input: DescribeBudgetPerformanceHistoryRequest
) -> RusotoFuture<DescribeBudgetPerformanceHistoryResponse, DescribeBudgetPerformanceHistoryError>

Describes the history for DAILY, MONTHLY, and QUARTERLY budgets. Budget history isn't available for ANNUAL budgets.

fn describe_budgets(
&self,
input: DescribeBudgetsRequest
) -> RusotoFuture<DescribeBudgetsResponse, DescribeBudgetsError>

Lists the budgets that are associated with an account.

The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

fn describe_notifications_for_budget(
&self,
input: DescribeNotificationsForBudgetRequest
) -> RusotoFuture<DescribeNotificationsForBudgetResponse, DescribeNotificationsForBudgetError>

Lists the notifications that are associated with a budget.

fn describe_subscribers_for_notification(
&self,
input: DescribeSubscribersForNotificationRequest
) -> RusotoFuture<DescribeSubscribersForNotificationResponse, DescribeSubscribersForNotificationError>

Lists the subscribers that are associated with a notification.

fn update_budget(
&self,
input: UpdateBudgetRequest
) -> RusotoFuture<UpdateBudgetResponse, UpdateBudgetError>

Updates a budget. You can change every part of a budget except for the budgetName and the calculatedSpend. When you modify a budget, the calculatedSpend drops to zero until AWS has new usage data to use for forecasting.

Only one of BudgetLimit or PlannedBudgetLimits can be present in the syntax at one time. Use the syntax that matches your case. The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

fn update_notification(
&self,
input: UpdateNotificationRequest
) -> RusotoFuture<UpdateNotificationResponse, UpdateNotificationError>

Updates a notification.

fn update_subscriber(
&self,
input: UpdateSubscriberRequest
) -> RusotoFuture<UpdateSubscriberResponse, UpdateSubscriberError>

Updates a subscriber.

Loading content...