Crate rusoto_budgets

The AWS Budgets API enables you to use AWS Budgets to plan your service usage, service costs, and instance reservations. The API reference provides descriptions, syntax, and usage examples for each of the actions and data types for AWS Budgets.

Budgets provide you with a way to see the following information:

How close your plan is to your budgeted amount or to the free tier limits

Your usage-to-date, including how much you've used of your Reserved Instances (RIs)

Your current estimated charges from AWS, and how much your predicted usage will accrue in charges by the end of the month

How much of your budget has been used

AWS updates your budget status several times a day. Budgets track your unblended costs, subscriptions, refunds, and RIs. You can create the following types of budgets:

Cost budgets - Plan how much you want to spend on a service.

Usage budgets - Plan how much you want to use one or more services.

RI utilization budgets - Define a utilization threshold, and receive alerts when your RI usage falls below that threshold. This lets you see if your RIs are unused or under-utilized.

RI coverage budgets - Define a coverage threshold, and receive alerts when the number of your instance hours that are covered by RIs fall below that threshold. This lets you see how much of your instance usage is covered by a reservation.

Service Endpoint

The AWS Budgets API provides the following endpoint:

https://budgets.amazonaws.com

For information about costs that are associated with the AWS Budgets API, see AWS Cost Management Pricing.

If you're using the service, you're probably looking for BudgetsClient and Budgets.

Structs

Budget
Represents the output of the CreateBudget operation. The content consists of the detailed metadata and data file information, and the current status of the budget object.

This is the ARN pattern for a budget:

arn:aws:budgetservice::AccountId:budget/budgetName

BudgetPerformanceHistory
A history of the state of a budget at the end of the budget's specified time period.

BudgetedAndActualAmounts
The amount of cost or usage that you created the budget for, compared to your actual costs or usage.

BudgetsClient
A client for the AWSBudgets API.

CalculatedSpend
The spend objects that are associated with this budget. The actualSpend tracks how much you've used, cost, usage, or RI units, and the forecastedSpend tracks how much you are predicted to spend if your current usage remains steady.

For example, if it is the 20th of the month and you have spent 50 dollars on Amazon EC2, your actualSpend is 50 USD, and your forecastedSpend is 75 USD.

CostTypes
The types of cost that are included in a COST budget, such as tax and subscriptions.

USAGE, RI_UTILIZATION, and RI_COVERAGE budgets do not have CostTypes.

CreateBudgetRequest
Request of CreateBudget

CreateBudgetResponse
Response of CreateBudget

CreateNotificationRequest
Request of CreateNotification

CreateNotificationResponse
Response of CreateNotification

CreateSubscriberRequest
Request of CreateSubscriber

CreateSubscriberResponse
Response of CreateSubscriber

DeleteBudgetRequest
Request of DeleteBudget

DeleteBudgetResponse
Response of DeleteBudget

DeleteNotificationRequest
Request of DeleteNotification

DeleteNotificationResponse
Response of DeleteNotification

DeleteSubscriberRequest
Request of DeleteSubscriber

DeleteSubscriberResponse
Response of DeleteSubscriber

DescribeBudgetPerformanceHistoryRequest

DescribeBudgetPerformanceHistoryResponse

DescribeBudgetRequest
Request of DescribeBudget

DescribeBudgetResponse
Response of DescribeBudget

DescribeBudgetsRequest
Request of DescribeBudgets

DescribeBudgetsResponse
Response of DescribeBudgets

DescribeNotificationsForBudgetRequest
Request of DescribeNotificationsForBudget

DescribeNotificationsForBudgetResponse
Response of GetNotificationsForBudget

DescribeSubscribersForNotificationRequest
Request of DescribeSubscribersForNotification

DescribeSubscribersForNotificationResponse
Response of DescribeSubscribersForNotification

Notification

A notification that is associated with a budget. A budget can have up to five notifications.

Each notification must have at least one subscriber. A notification can have one SNS subscriber and up to 10 email subscribers, for a total of 11 subscribers.

For example, if you have a budget for 200 dollars and you want to be notified when you go over 160 dollars, create a notification with the following parameters:

A notificationType of ACTUAL

A thresholdType of PERCENTAGE

A comparisonOperator of GREATER_THAN

A notification threshold of 80

NotificationWithSubscribers
A notification with subscribers. A notification can have one SNS subscriber and up to 10 email subscribers, for a total of 11 subscribers.

Spend

The amount of cost or usage that is measured for a budget.

For example, a Spend for 3 GB of S3 usage would have the following parameters:

An Amount of 3

A unit of GB

Subscriber

The subscriber to a budget notification. The subscriber consists of a subscription type and either an Amazon SNS topic or an email address.

For example, an email subscriber would have the following parameters:

A subscriptionType of EMAIL

An address of example@example.com

TimePeriod
The period of time that is covered by a budget. The period has a start date and an end date. The start date must come before the end date. There are no restrictions on the end date.

UpdateBudgetRequest
Request of UpdateBudget

UpdateBudgetResponse
Response of UpdateBudget

UpdateNotificationRequest
Request of UpdateNotification

UpdateNotificationResponse
Response of UpdateNotification

UpdateSubscriberRequest
Request of UpdateSubscriber

UpdateSubscriberResponse
Response of UpdateSubscriber

Enums

CreateBudgetError
Errors returned by CreateBudget

CreateNotificationError
Errors returned by CreateNotification

CreateSubscriberError
Errors returned by CreateSubscriber

DeleteBudgetError
Errors returned by DeleteBudget

DeleteNotificationError
Errors returned by DeleteNotification

DeleteSubscriberError
Errors returned by DeleteSubscriber

DescribeBudgetError
Errors returned by DescribeBudget

DescribeBudgetPerformanceHistoryError
Errors returned by DescribeBudgetPerformanceHistory

DescribeBudgetsError
Errors returned by DescribeBudgets

DescribeNotificationsForBudgetError
Errors returned by DescribeNotificationsForBudget

DescribeSubscribersForNotificationError
Errors returned by DescribeSubscribersForNotification

UpdateBudgetError
Errors returned by UpdateBudget

UpdateNotificationError
Errors returned by UpdateNotification

UpdateSubscriberError
Errors returned by UpdateSubscriber

Traits

Budgets
Trait representing the capabilities of the AWSBudgets API. AWSBudgets clients implement this trait.