Crate rusoto_budgets

Budgets enable you to plan your service usage, service costs, and your RI utilization. You can also track how close your plan is to your budgeted amount or to the free tier limits. Budgets provide you with a quick way to see your usage-to-date and current estimated charges from AWS and to see how much your predicted usage accrues in charges by the end of the month. Budgets also compare current estimates and charges to the amount that you indicated you want to use or spend and lets you see how much of your budget has been used. AWS updates your budget status several times a day. Budgets track your unblended costs, subscriptions, and refunds. You can create the following types of budgets:

Cost budgets allow you to say how much you want to spend on a service.

Usage budgets allow you to say how many hours you want to use for one or more services.

RI utilization budgets allow you to define a utilization threshold and receive alerts when RIs are tracking below that threshold.

You can create up to 20,000 budgets per AWS master account. Your first two budgets are free of charge. Each additional budget costs $0.02 per day. You can set up optional notifications that warn you if you exceed, or are forecasted to exceed, your budgeted amount. You can have notifications sent to an Amazon SNS topic, to an email address, or to both. For more information, see Creating an Amazon SNS Topic for Budget Notifications. AWS Free Tier usage alerts via AWS Budgets are provided for you, and do not count toward your budget limits.

Service Endpoint

The AWS Budgets API provides the following endpoint:

https://budgets.amazonaws.com

For information about costs associated with the AWS Budgets API, see AWS Cost Management Pricing.

If you're using the service, you're probably looking for BudgetsClient and Budgets.

Structs

Budget
Represents the output of the CreateBudget operation. The content consists of the detailed metadata and data file information, and the current status of the budget.

The ARN pattern for a budget is: arn:aws:budgetservice::AccountId:budget/budgetName

BudgetsClient
A client for the AWSBudgets API.

CalculatedSpend
The spend objects associated with this budget. The actualSpend tracks how much you've used, cost, usage, or RI units, and the forecastedSpend tracks how much you are predicted to spend if your current usage remains steady.

For example, if it is the 20th of the month and you have spent 50 dollars on Amazon EC2, your actualSpend is 50 USD, and your forecastedSpend is 75 USD.

CostTypes
The types of cost included in a budget, such as tax and subscriptions.

CreateBudgetRequest
Request of CreateBudget

CreateBudgetResponse
Response of CreateBudget

CreateNotificationRequest
Request of CreateNotification

CreateNotificationResponse
Response of CreateNotification

CreateSubscriberRequest
Request of CreateSubscriber

CreateSubscriberResponse
Response of CreateSubscriber

DeleteBudgetRequest
Request of DeleteBudget

DeleteBudgetResponse
Response of DeleteBudget

DeleteNotificationRequest
Request of DeleteNotification

DeleteNotificationResponse
Response of DeleteNotification

DeleteSubscriberRequest
Request of DeleteSubscriber

DeleteSubscriberResponse
Response of DeleteSubscriber

DescribeBudgetRequest
Request of DescribeBudget

DescribeBudgetResponse
Response of DescribeBudget

DescribeBudgetsRequest
Request of DescribeBudgets

DescribeBudgetsResponse
Response of DescribeBudgets

DescribeNotificationsForBudgetRequest
Request of DescribeNotificationsForBudget

DescribeNotificationsForBudgetResponse
Response of GetNotificationsForBudget

DescribeSubscribersForNotificationRequest
Request of DescribeSubscribersForNotification

DescribeSubscribersForNotificationResponse
Response of DescribeSubscribersForNotification

Notification

A notification associated with a budget. A budget can have up to five notifications.

Each notification must have at least one subscriber. A notification can have one SNS subscriber and up to ten email subscribers, for a total of 11 subscribers.

For example, if you have a budget for 200 dollars and you want to be notified when you go over 160 dollars, create a notification with the following parameters:

A notificationType of ACTUAL

A comparisonOperator of GREATER_THAN

A notification threshold of 80

NotificationWithSubscribers
A notification with subscribers. A notification can have one SNS subscriber and up to ten email subscribers, for a total of 11 subscribers.

Spend

The amount of cost or usage being measured for a budget.

For example, a Spend for 3 GB of S3 usage would have the following parameters:

An Amount of 3

A unit of GB

Subscriber

The subscriber to a budget notification. The subscriber consists of a subscription type and either an Amazon Simple Notification Service topic or an email address.

For example, an email subscriber would have the following parameters:

A subscriptionType of EMAIL

An address of example@example.com

TimePeriod
The period of time covered by a budget. Has a start date and an end date. The start date must come before the end date. There are no restrictions on the end date.

UpdateBudgetRequest
Request of UpdateBudget

UpdateBudgetResponse
Response of UpdateBudget

UpdateNotificationRequest
Request of UpdateNotification

UpdateNotificationResponse
Response of UpdateNotification

UpdateSubscriberRequest
Request of UpdateSubscriber

UpdateSubscriberResponse
Response of UpdateSubscriber

Enums

CreateBudgetError
Errors returned by CreateBudget

CreateNotificationError
Errors returned by CreateNotification

CreateSubscriberError
Errors returned by CreateSubscriber

DeleteBudgetError
Errors returned by DeleteBudget

DeleteNotificationError
Errors returned by DeleteNotification

DeleteSubscriberError
Errors returned by DeleteSubscriber

DescribeBudgetError
Errors returned by DescribeBudget

DescribeBudgetsError
Errors returned by DescribeBudgets

DescribeNotificationsForBudgetError
Errors returned by DescribeNotificationsForBudget

DescribeSubscribersForNotificationError
Errors returned by DescribeSubscribersForNotification

UpdateBudgetError
Errors returned by UpdateBudget

UpdateNotificationError
Errors returned by UpdateNotification

UpdateSubscriberError
Errors returned by UpdateSubscriber

Traits

Budgets
Trait representing the capabilities of the AWSBudgets API. AWSBudgets clients implement this trait.